refactored ComponentIFrame and added default values for props
